+++ lib/src/classes/images/guild_discovery_splash.dart
@@ -1,0 +1,10 @@
+class GuildDiscoverySplash  {
+  late String id;
+  late String hash;
+
+  GuildDiscoverySplash(this.hash, this.id);
+
+  String url({String? extension = "jpeg"}) {
+    return "https://cdn.discordapp.com/discovery-splashes/$id/$hash.$extension";
+  }
+}