+++ lib/src/main.dart
@@ -20,7 +20,7 @@ final apiURL = "https://discord.com/api/v$version";
 /// This function calculate the intent number required from the gateway.
 /// [intents] is a list of multiples of two. You can use GatewayIntentBits class.
 /// Return a number.
-int calculateIntents(List<int> intents) {
+int intentsCalculator(List<int> intents) {
   int intentsNumber = 0;
 
   for (var element in intents) {
@@ -138,9 +138,6 @@ class Client extends EventEmitter {
         case "GUILD_CREATE":
           if (guilds.cache.has(event["d"]["id"])) {
             Guild oldGuild = guilds.cache.get(event["d"]["id"]);
-            if (oldGuild.appIsOwner != null) {
-              event['d']["owner"] = oldGuild.appIsOwner;
-            }
             if (oldGuild.permissions != null) {
               event['d']["permissions"] = oldGuild.permissions;
             }
@@ -156,13 +153,11 @@ class Client extends EventEmitter {
           }
           break;
         case "GUILD_DELETE":
-          dynamic guild;
           if (guilds.cache.has(event["d"]["id"])) {
-            guild = guilds.cache.get(event["d"]["id"]);
-            guilds.cache.set(event["d"]["id"], UnavailableGuild(event["d"]["id"], notUpdatedGuild: guild));
-          } else {
-            guild = event["d"];
+            guilds.cache.delete(event["d"]["id"]);
+            guilds.cache.set(event["d"]["id"], UnavailableGuild(event["d"]["id"],));
           }
+          var guild = event["d"];
           emit("GUILD_DELETE", guild);
           break;
         case "INTERACTION_CREATE":